requirements.txt
tensorflow-cpu==2.16.1
tf-keras==2.16.0

Keras-Preprocessing==1.1.2

spacy>=3.5.0
sentence-transformers>=5.1.0

datasets>=4.1.1
pandas>=2.0.0
numpy>=1.25.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.12.2

wordcloud>=1.9.4
nltk>=3.9.1

fastapi>=0.96.0
python-multipart>=0.0.6
supabase>=2.20.0

streamlit>=1.50.0

python-dotenv>=1.0.0
pydantic-settings>=2.0.0
shap>=0.42.1
uniconv

ml-dtypes==0.3.1
protobuf>=3.20.3,<3.21
tensorboard==2.16
